requirements.txt
pandas
os
ydata_profiling 
requests
hashlib
zipfile
matplotlib.pyplot 
scikit-learn
snakemake
numpy
